RecordingPreStartModel, RecordingPrepRecFailedModel のコマンド実行処理を ReserveExternalProcessModel へ集約
