+++ src/server/Model/Operator/Callbacks/RecordingPrepRecFailedModel.ts
@@ -1,8 +1,7 @@
-import { spawn } from 'child_process';
-import ProcessUtil from '../../../Util/ProcessUtil';
 import * as DBSchema from '../../DB/DBSchema';
 import Model from '../../Model';
 import { RecordingManageModelInterface } from '../../Operator/Recording/RecordingManageModel';
+import { ReserveExternalProcessModelInterface } from '../../Operator/ReserveExternalProcessModel';
 import CallbackBaseModelInterface from './CallbackBaseModelInterface';
 
 /**
@@ -11,13 +10,16 @@ import CallbackBaseModelInterface from './CallbackBaseModelInterface';
  */
 class RecordingPrepRecFailedModel extends Model implements CallbackBaseModelInterface {
     private recordingManage: RecordingManageModelInterface;
+    private externalProcess: ReserveExternalProcessModelInterface;
 
     constructor(
         recordingManage: RecordingManageModelInterface,
+        externalProcess: ReserveExternalProcessModelInterface,
     ) {
         super();
 
         this.recordingManage = recordingManage;
+        this.externalProcess = externalProcess;
     }
 
     public set(): void {
@@ -25,46 +27,14 @@ class RecordingPrepRecFailedModel extends Model implements CallbackBaseModelInte
     }
 
     /**
-     * @param program: DBSchema.RecordedSchema
+     * @param program: DBSchema.ProgramSchema
      */
     private callback(program: DBSchema.ProgramSchema): void {
         // 外部コマンド実行
         const cmd = this.config.getConfig().recordedPrepRecFailedCommand;
-        if (typeof cmd === 'undefined') { return; }
-
-        this.log.system.info(`run: ${ cmd }`);
-
-        let cmds: ProcessUtil.Cmds;
-        try {
-            cmds = ProcessUtil.parseCmdStr(cmd);
-        } catch (err) {
-            this.log.system.error(<any> err);
-
-            return;
+        if (typeof cmd !== 'undefined') {
+            this.externalProcess.run(cmd, program, 'recording preprec failed');
         }
-
-        const child = spawn(cmds.bin, cmds.args, {
-            env: {
-                PROGRAMID: program.id,
-                CHANNELTYPE: program.channelType,
-                CHANNELID: program.channelId,
-                STARTAT: program.startAt,
-                ENDAT: program.endAt,
-                DURATION: program.duration,
-                NAME: program.name,
-                DESCRIPTION: program.description,
-                EXTENDED: program.extended,
-            },
-        });
-
-        child.on('exit', () => {
-            this.log.system.info('recording preprec failed process is fin');
-        });
-
-        child.on('error', (err) => {
-            this.log.system.error('recording preprec failed process is error');
-            this.log.system.error(String(err));
-        });
     }
 }
